refactor: decompose high-churn Compose screens into focused components (Phase C-1)
Extracted from five top-churn screens into per-module `components/` packages.
Pure UI code motion — no behavior, visual, or ViewModel changes.

rider-app:
  - HistoryScreen → HistoryList, HistoryFilterBar, HistoryEntryCard, HistoryStatsCard
  - RiderModeScreen → RideRequestPanel, ActiveRidePanel, CompletionPanel, CancelDialogStack

drivestr:
  - DriverModeScreen → AvailabilityControls, OfferInbox, RoadflareFollowerList,
    RideOfferCard, BroadcastRideRequestCard, NoCommonPaymentMethodDialog,
    formatRelativeTime/Distance/Earnings helpers

roadflare-rider:
  - DriverNetworkTab → FollowedDriverList, DriverCard, DriverStatusBadge,
    EmptyDriversState, RoadflarePaymentMethodsDialog
  - HistoryScreen → HistoryList, HistoryFilterBar, HistoryEntryCard, HistoryStatsCard
  - RideTab (layout-stable only) → CancelledContent, formatFareAmount
    Adds TODO header for deferred state-binding extractions blocked on #65.
