+++ drivestr/src/main/java/com/drivestr/app/ui/screens/components/DriverModeScreenComponents.kt
@@ -1,0 +1,1030 @@
+package com.drivestr.app.ui.screens.components
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.*
+import androidx.compose.material3.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import com.ridestr.common.bitcoin.BitcoinPriceService
+import com.ridestr.common.fiat.usdAmountOrNull
+import com.ridestr.common.location.GeocodingService
+import com.ridestr.common.nostr.events.BroadcastRideOfferData
+import com.ridestr.common.nostr.events.RideOfferData
+import com.ridestr.common.routing.RouteResult
+import com.ridestr.common.settings.DisplayCurrency
+import com.ridestr.common.settings.DistanceUnit
+import com.ridestr.common.ui.FareDisplay
+
+// ---------------------------------------------------------------------------
+// Utility functions
+// ---------------------------------------------------------------------------
+
+/**
+ * Format a Unix timestamp as relative time (e.g., "5s ago", "2m ago").
+ */
+fun formatRelativeTime(timestampSeconds: Long): String {
+    val nowSeconds = System.currentTimeMillis() / 1000
+    val diffSeconds = nowSeconds - timestampSeconds
+
+    return when {
+        diffSeconds < 60 -> "${diffSeconds}s ago"
+        diffSeconds < 3600 -> "${diffSeconds / 60}m ago"
+        diffSeconds < 86400 -> "${diffSeconds / 3600}h ago"
+        else -> "${diffSeconds / 86400}d ago"
+    }
+}
+
+/**
+ * Format distance based on user's preferred unit.
+ */
+fun formatDistance(distanceKm: Double, unit: DistanceUnit): String {
+    return when (unit) {
+        DistanceUnit.MILES -> {
+            val miles = distanceKm * 0.621371
+            if (miles < 0.1) {
+                "${(miles * 5280).toInt()} ft"
+            } else {
+                String.format("%.1f mi", miles)
+            }
+        }
+        DistanceUnit.KILOMETERS -> {
+            if (distanceKm < 1) {
+                "${(distanceKm * 1000).toInt()} m"
+            } else {
+                String.format("%.1f km", distanceKm)
+            }
+        }
+    }
+}
+
+/**
+ * Format earnings rate (sats/unit) with currency conversion.
+ */
+fun formatEarnings(
+    satsPerUnit: Double,
+    displayCurrency: DisplayCurrency,
+    btcPriceUsd: Int?,
+    suffix: String,
+    fiatAmountPerUnitUsd: Double? = null
+): String {
+    return when (displayCurrency) {
+        DisplayCurrency.SATS -> {
+            "${satsPerUnit.toInt()} sats$suffix"
+        }
+        DisplayCurrency.USD -> {
+            if (fiatAmountPerUnitUsd != null) {
+                String.format("$%.2f$suffix", fiatAmountPerUnitUsd)
+            } else if (btcPriceUsd != null && btcPriceUsd > 0) {
+                // Convert sats to USD: sats / 100_000_000 * btcPriceUsd
+                val usdValue = (satsPerUnit / 100_000_000.0) * btcPriceUsd
+                String.format("$%.2f$suffix", usdValue)
+            } else {
+                "${satsPerUnit.toInt()} sats$suffix"
+            }
+        }
+    }
+}
+
+fun effectiveRoadflareDriverMethods(driverFiatMethods: List<String>): List<String> {
+    val bitcoin = com.ridestr.common.nostr.events.PaymentMethod.BITCOIN.value
+    return (driverFiatMethods + bitcoin).distinctBy { it.trim().lowercase() }
+}
+
+// ---------------------------------------------------------------------------
+// Extracted composables: moved from DriverModeScreen
+// ---------------------------------------------------------------------------
+
+@Composable
+fun RideOfferCard(
+    offer: RideOfferData,
+    pickupRoute: RouteResult?,
+    rideRoute: RouteResult?,
+    isProcessing: Boolean,
+    onAccept: () -> Unit,
+    onDecline: () -> Unit,
+    displayCurrency: DisplayCurrency,
+    distanceUnit: DistanceUnit,
+    onToggleCurrency: () -> Unit,
+    priceService: BitcoinPriceService,
+    driverFiatMethods: List<String>
+) {
+    val context = LocalContext.current
+    val geocodingService = remember { GeocodingService(context) }
+    val btcPrice by priceService.btcPriceUsd.collectAsState()
+
+    // Update relative time every second
+    var relativeTime by remember { mutableStateOf(formatRelativeTime(offer.createdAt)) }
+    LaunchedEffect(offer.createdAt) {
+        while (true) {
+            relativeTime = formatRelativeTime(offer.createdAt)
+            kotlinx.coroutines.delay(1000)
+        }
+    }
+
+    // Geocode pickup and destination addresses
+    var pickupAddress by remember { mutableStateOf<String?>(null) }
+    var destinationAddress by remember { mutableStateOf<String?>(null) }
+
+    LaunchedEffect(offer.approxPickup, offer.destination) {
+        pickupAddress = geocodingService.reverseGeocode(offer.approxPickup.lat, offer.approxPickup.lon)
+        destinationAddress = geocodingService.reverseGeocode(offer.destination.lat, offer.destination.lon)
+    }
+
+    // Route info calculations
+    val pickupDistanceKm = pickupRoute?.distanceKm
+    val pickupDurationMin = pickupRoute?.let { it.durationSeconds / 60.0 }
+    val rideDistanceKm = rideRoute?.distanceKm
+    val rideDurationMin = rideRoute?.let { it.durationSeconds / 60.0 }
+
+    // Calculate earnings metrics (only if we have both routes)
+    val earningsPerHour: String?
+    val earningsPerDistance: String?
+    val authoritativeUsdFare = offer.fiatFare?.usdAmountOrNull()
+
+    if (pickupRoute != null && rideRoute != null && pickupDurationMin != null && rideDurationMin != null) {
+        val totalTimeHours = (pickupDurationMin + rideDurationMin) / 60.0
+        val totalDistanceKm = (pickupDistanceKm ?: 0.0) + (rideDistanceKm ?: 0.0)
+        val totalDistanceForEarnings = if (distanceUnit == DistanceUnit.MILES) {
+            totalDistanceKm * 0.621371
+        } else {
+            totalDistanceKm
+        }
+
+        earningsPerHour = if (totalTimeHours > 0) {
+            formatEarnings(
+                satsPerUnit = offer.fareEstimate / totalTimeHours,
+                displayCurrency = displayCurrency,
+                btcPriceUsd = btcPrice,
+                suffix = "/hr",
+                fiatAmountPerUnitUsd = authoritativeUsdFare?.div(totalTimeHours)
+            )
+        } else null
+
+        earningsPerDistance = if (totalDistanceForEarnings > 0) {
+            val perDistanceUnit = if (distanceUnit == DistanceUnit.MILES) "/mi" else "/km"
+            formatEarnings(
+                satsPerUnit = offer.fareEstimate / totalDistanceForEarnings,
+                displayCurrency = displayCurrency,
+                btcPriceUsd = btcPrice,
+                suffix = perDistanceUnit,
+                fiatAmountPerUnitUsd = authoritativeUsdFare?.div(totalDistanceForEarnings)
+            )
+        } else null
+    } else {
+        earningsPerHour = null
+        earningsPerDistance = null
+    }
+
+    Card(modifier = Modifier.fillMaxWidth()) {
+        Column(modifier = Modifier.padding(16.dp)) {
+            // Header: Fare + Time
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Column {
+                    Row(verticalAlignment = Alignment.CenterVertically) {
+                        Icon(
+                            Icons.Default.Lock,
+                            contentDescription = null,
+                            modifier = Modifier.size(20.dp),
+                            tint = MaterialTheme.colorScheme.primary
+                        )
+                        Spacer(modifier = Modifier.width(4.dp))
+                        FareDisplay(
+                            satsAmount = offer.fareEstimate,
+                            displayCurrency = displayCurrency,
+                            onToggleCurrency = onToggleCurrency,
+                            priceService = priceService,
+                            fiatFare = offer.fiatFare,
+                            style = MaterialTheme.typography.headlineSmall
+                        )
+                    }
+                    Text(
+                        text = if (offer.isRoadflare) "RoadFlare" else "Direct Request",
+                        style = MaterialTheme.typography.labelSmall,
+                        color = if (offer.isRoadflare) MaterialTheme.colorScheme.tertiary else MaterialTheme.colorScheme.primary
+                    )
+                    // Show fiat payment match status for RoadFlare offers (Issue #46)
+                    if (offer.isRoadflare && offer.paymentMethod != "cashu") {
+                        val bestMatch = if (offer.fiatPaymentMethods.isNotEmpty()) {
+                            com.ridestr.common.nostr.events.PaymentMethod.findBestCommonFiatMethod(
+                                offer.fiatPaymentMethods,
+                                effectiveRoadflareDriverMethods(driverFiatMethods)
+                            )
+                        } else null
+                        val methodDisplay = com.ridestr.common.nostr.events.PaymentMethod.fromString(offer.paymentMethod)?.displayName ?: offer.paymentMethod
+
+                        if (bestMatch != null) {
+                            val matchDisplay = com.ridestr.common.nostr.events.PaymentMethod.fromString(bestMatch)?.displayName ?: bestMatch
+                            Text(
+                                text = "Pay via: $matchDisplay",
+                                style = MaterialTheme.typography.labelSmall,
+                                color = androidx.compose.ui.graphics.Color(0xFF2E7D32) // Green
+                            )
+                        } else if (offer.fiatPaymentMethods.isNotEmpty()) {
+                            Text(
+                                text = "No Common Payment Method",
+                                style = MaterialTheme.typography.labelSmall,
+                                color = MaterialTheme.colorScheme.error
+                            )
+                        } else {
+                            // Legacy offer without fiat_payment_methods field
+                            Text(
+                                text = "Payment: $methodDisplay",
+                                style = MaterialTheme.typography.labelSmall,
+                                color = MaterialTheme.colorScheme.tertiary
+                            )
+                        }
+                    }
+                }
+                Text(
+                    text = relativeTime,
+                    style = MaterialTheme.typography.bodySmall,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+            }
+
+            Spacer(modifier = Modifier.height(12.dp))
+
+            // Pickup route info
+            Row(verticalAlignment = Alignment.CenterVertically) {
+                Icon(
+                    Icons.Default.DirectionsCar,
+                    contentDescription = null,
+                    modifier = Modifier.size(18.dp),
+                    tint = MaterialTheme.colorScheme.primary
+                )
+                Spacer(modifier = Modifier.width(6.dp))
+                if (pickupDistanceKm != null && pickupDurationMin != null) {
+                    Text(
+                        text = "${formatDistance(pickupDistanceKm, distanceUnit)} away",
+                        style = MaterialTheme.typography.bodyMedium,
+                        fontWeight = FontWeight.Medium
+                    )
+                    Text(
+                        text = " • ${pickupDurationMin.toInt()} min to pickup",
+                        style = MaterialTheme.typography.bodyMedium,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                } else {
+                    Text(
+                        text = "Calculating route...",
+                        style = MaterialTheme.typography.bodyMedium,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+            }
+
+            Spacer(modifier = Modifier.height(6.dp))
+
+            // Ride route info
+            Row(verticalAlignment = Alignment.CenterVertically) {
+                Icon(
+                    Icons.Default.Place,
+                    contentDescription = null,
+                    modifier = Modifier.size(18.dp),
+                    tint = MaterialTheme.colorScheme.tertiary
+                )
+                Spacer(modifier = Modifier.width(6.dp))
+                if (rideDistanceKm != null && rideDurationMin != null) {
+                    Text(
+                        text = "${formatDistance(rideDistanceKm, distanceUnit)} ride",
+                        style = MaterialTheme.typography.bodyMedium,
+                        fontWeight = FontWeight.Medium
+                    )
+                    Text(
+                        text = " • ${rideDurationMin.toInt()} min",
+                        style = MaterialTheme.typography.bodyMedium,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                } else {
+                    Text(
+                        text = "Calculating...",
+                        style = MaterialTheme.typography.bodyMedium,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+            }
+
+            // Earnings metrics (if available)
+            if (earningsPerHour != null || earningsPerDistance != null) {
+                Spacer(modifier = Modifier.height(8.dp))
+                Row(
+                    horizontalArrangement = Arrangement.spacedBy(16.dp)
+                ) {
+                    earningsPerHour?.let {
+                        Text(
+                            text = it,
+                            style = MaterialTheme.typography.labelMedium,
+                            color = MaterialTheme.colorScheme.primary
+                        )
+                    }
+                    earningsPerDistance?.let {
+                        Text(
+                            text = it,
+                            style = MaterialTheme.typography.labelMedium,
+                            color = MaterialTheme.colorScheme.primary
+                        )
+                    }
+                }
+            }
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.spacedBy(8.dp)
+            ) {
+                OutlinedButton(
+                    onClick = onDecline,
+                    enabled = !isProcessing,
+                    modifier = Modifier.weight(1f),
+                    colors = ButtonDefaults.outlinedButtonColors(
+                        contentColor = MaterialTheme.colorScheme.error
+                    )
+                ) {
+                    Icon(Icons.Default.Close, contentDescription = null, modifier = Modifier.size(18.dp))
+                    Spacer(modifier = Modifier.width(4.dp))
+                    Text("Decline")
+                }
+                Button(
+                    onClick = onAccept,
+                    enabled = !isProcessing,
+                    modifier = Modifier.weight(1f)
+                ) {
+                    if (isProcessing) {
+                        CircularProgressIndicator(
+                            modifier = Modifier.size(18.dp),
+                            strokeWidth = 2.dp
+                        )
+                    } else {
+                        Icon(Icons.Default.Check, contentDescription = null, modifier = Modifier.size(18.dp))
+                        Spacer(modifier = Modifier.width(4.dp))
+                        Text("Accept")
+                    }
+                }
+            }
+        }
+    }
+}
+
+/**
+ * Card for displaying a broadcast ride request.
+ * Shows fare prominently, route info, earnings metrics, and request age.
+ */
+@Composable
+fun BroadcastRideRequestCard(
+    request: BroadcastRideOfferData,
+    pickupRoute: RouteResult?,
+    isProcessing: Boolean,
+    onAccept: () -> Unit,
+    onDecline: () -> Unit,
+    displayCurrency: DisplayCurrency,
+    distanceUnit: DistanceUnit,
+    onToggleCurrency: () -> Unit,
+    priceService: BitcoinPriceService
+) {
+    val btcPrice by priceService.btcPriceUsd.collectAsState()
+
+    // Update relative time every second
+    var relativeTime by remember { mutableStateOf(formatRelativeTime(request.createdAt)) }
+    LaunchedEffect(request.createdAt) {
+        while (true) {
+            relativeTime = formatRelativeTime(request.createdAt)
+            kotlinx.coroutines.delay(1000)
+        }
+    }
+
+    // Distance conversion (km to miles if needed)
+    val rideDistanceKm = request.routeDistanceKm
+    val rideDurationMin = request.routeDurationMin
+
+    // Format ride distance based on user preference
+    val rideDistanceStr = formatDistance(rideDistanceKm, distanceUnit)
+
+    // Pickup route info (only available if calculated)
+    val pickupDistanceKm = pickupRoute?.distanceKm
+    val pickupDurationMin = pickupRoute?.let { it.durationSeconds / 60.0 }
+    val pickupDistanceStr = pickupDistanceKm?.let { formatDistance(it, distanceUnit) }
+
+    // Calculate earnings metrics ONLY if we have pickup route data
+    // Without pickup time, $/hr would be misleadingly high
+    val earningsPerHour: String?
+    val earningsPerDistance: String?
+    val authoritativeUsdFare = request.fiatFare?.usdAmountOrNull()
+
+    if (pickupRoute != null && pickupDurationMin != null && pickupDistanceKm != null) {
+        // Total time = pickup time + ride time (in hours)
+        val totalTimeHours = (pickupDurationMin + rideDurationMin) / 60.0
+
+        // Total distance = pickup + ride (driver has to drive both, only gets paid for ride)
+        // This gives accurate $/mile for total miles driven
+        val totalDistanceKm = pickupDistanceKm + rideDistanceKm
+        val totalDistanceForEarnings = if (distanceUnit == DistanceUnit.MILES) {
+            totalDistanceKm * 0.621371
+        } else {
+            totalDistanceKm
+        }
+
+        earningsPerHour = if (totalTimeHours > 0) {
+            formatEarnings(
+                satsPerUnit = request.fareEstimate / totalTimeHours,
+                displayCurrency = displayCurrency,
+                btcPriceUsd = btcPrice,
+                suffix = "/hr",
+                fiatAmountPerUnitUsd = authoritativeUsdFare?.div(totalTimeHours)
+            )
+        } else null
+
+        earningsPerDistance = if (totalDistanceForEarnings > 0) {
+            val perDistanceUnit = if (distanceUnit == DistanceUnit.MILES) "/mi" else "/km"
+            formatEarnings(
+                satsPerUnit = request.fareEstimate / totalDistanceForEarnings,
+                displayCurrency = displayCurrency,
+                btcPriceUsd = btcPrice,
+                suffix = perDistanceUnit,
+                fiatAmountPerUnitUsd = authoritativeUsdFare?.div(totalDistanceForEarnings)
+            )
+        } else null
+    } else {
+        // Don't show earnings without pickup route - they'd be misleading
+        earningsPerHour = null
+        earningsPerDistance = null
+    }
+
+    Card(
+        modifier = Modifier.fillMaxWidth(),
+        colors = CardDefaults.cardColors(
+            containerColor = MaterialTheme.colorScheme.surface
+        )
+    ) {
+        Column(modifier = Modifier.padding(16.dp)) {
+            // Header: Fare + Time
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Column {
+                    Row(verticalAlignment = Alignment.CenterVertically) {
+                        Icon(
+                            Icons.Default.Bolt,
+                            contentDescription = null,
+                            modifier = Modifier.size(24.dp),
+                            tint = MaterialTheme.colorScheme.primary
+                        )
+                        Spacer(modifier = Modifier.width(4.dp))
+                        FareDisplay(
+                            satsAmount = request.fareEstimate,
+                            displayCurrency = displayCurrency,
+                            onToggleCurrency = onToggleCurrency,
+                            priceService = priceService,
+                            fiatFare = request.fiatFare,
+                            style = MaterialTheme.typography.headlineSmall
+                        )
+                    }
+                    if (request.isRoadflare) {
+                        Text(
+                            text = "RoadFlare",
+                            style = MaterialTheme.typography.labelSmall,
+                            color = MaterialTheme.colorScheme.tertiary
+                        )
+                    }
+                }
+                Text(
+                    text = relativeTime,
+                    style = MaterialTheme.typography.bodySmall,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+            }
+
+            Spacer(modifier = Modifier.height(12.dp))
+
+            // Pickup info - distance/time from driver's location
+            Row(verticalAlignment = Alignment.CenterVertically) {
+                Icon(
+                    Icons.Default.DirectionsCar,
+                    contentDescription = null,
+                    modifier = Modifier.size(18.dp),
+                    tint = MaterialTheme.colorScheme.primary
+                )
+                Spacer(modifier = Modifier.width(6.dp))
+                if (pickupDistanceStr != null && pickupDurationMin != null) {
+                    Text(
+                        text = "$pickupDistanceStr away",
+                        style = MaterialTheme.typography.bodyMedium,
+                        fontWeight = FontWeight.Medium
+                    )
+                    Text(
+                        text = " • ${pickupDurationMin.toInt()} min to pickup",
+                        style = MaterialTheme.typography.bodyMedium,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                } else {
+                    Text(
+                        text = "Calculating route...",
+                        style = MaterialTheme.typography.bodyMedium,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+            }
+
+            Spacer(modifier = Modifier.height(6.dp))
+
+            // Ride info - distance/time of the trip
+            Row(verticalAlignment = Alignment.CenterVertically) {
+                Icon(
+                    Icons.Default.Place,
+                    contentDescription = null,
+                    modifier = Modifier.size(18.dp),
+                    tint = MaterialTheme.colorScheme.tertiary
+                )
+                Spacer(modifier = Modifier.width(6.dp))
+                Text(
+                    text = "$rideDistanceStr ride",
+                    style = MaterialTheme.typography.bodyMedium,
+                    fontWeight = FontWeight.Medium
+                )
+                Text(
+                    text = " • ${rideDurationMin.toInt()} min trip",
+                    style = MaterialTheme.typography.bodyMedium,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+            }
+
+            Spacer(modifier = Modifier.height(10.dp))
+
+            // Earnings metrics row
+            if (earningsPerHour != null || earningsPerDistance != null) {
+                Row(
+                    modifier = Modifier.fillMaxWidth(),
+                    horizontalArrangement = Arrangement.spacedBy(16.dp)
+                ) {
+                    // $/hr metric
+                    earningsPerHour?.let { perHour ->
+                        Row(verticalAlignment = Alignment.CenterVertically) {
+                            Icon(
+                                Icons.Default.Schedule,
+                                contentDescription = null,
+                                modifier = Modifier.size(14.dp),
+                                tint = MaterialTheme.colorScheme.secondary
+                            )
+                            Spacer(modifier = Modifier.width(4.dp))
+                            Text(
+                                text = perHour,
+                                style = MaterialTheme.typography.bodySmall,
+                                fontWeight = FontWeight.Medium,
+                                color = MaterialTheme.colorScheme.secondary
+                            )
+                        }
+                    }
+                    // $/mi or $/km metric
+                    earningsPerDistance?.let { perDistance ->
+                        Row(verticalAlignment = Alignment.CenterVertically) {
+                            Icon(
+                                Icons.Default.Straighten,
+                                contentDescription = null,
+                                modifier = Modifier.size(14.dp),
+                                tint = MaterialTheme.colorScheme.secondary
+                            )
+                            Spacer(modifier = Modifier.width(4.dp))
+                            Text(
+                                text = perDistance,
+                                style = MaterialTheme.typography.bodySmall,
+                                fontWeight = FontWeight.Medium,
+                                color = MaterialTheme.colorScheme.secondary
+                            )
+                        }
+                    }
+                }
+
+                Spacer(modifier = Modifier.height(12.dp))
+            }
+
+            // Action buttons
+            Row(
+                modifier = Modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.spacedBy(8.dp)
+            ) {
+                OutlinedButton(
+                    onClick = onDecline,
+                    enabled = !isProcessing,
+                    modifier = Modifier.weight(1f),
+                    colors = ButtonDefaults.outlinedButtonColors(
+                        contentColor = MaterialTheme.colorScheme.error
+                    )
+                ) {
+                    Icon(Icons.Default.Close, contentDescription = null, modifier = Modifier.size(18.dp))
+                    Spacer(modifier = Modifier.width(4.dp))
+                    Text("Pass")
+                }
+                Button(
+                    onClick = onAccept,
+                    enabled = !isProcessing,
+                    modifier = Modifier.weight(1f)
+                ) {
+                    if (isProcessing) {
+                        CircularProgressIndicator(
+                            modifier = Modifier.size(18.dp),
+                            strokeWidth = 2.dp
+                        )
+                    } else {
+                        Icon(Icons.Default.Check, contentDescription = null, modifier = Modifier.size(18.dp))
+                        Spacer(modifier = Modifier.width(4.dp))
+                        Text("Accept")
+                    }
+                }
+            }
+        }
+    }
+}
+
+/**
+ * Warning dialog shown when driver accepts a fiat offer with no common payment method.
+ * Shows rider's top preferences; "Decline Ride" closes dialog but keeps offer visible.
+ */
+@Composable
+fun NoCommonPaymentMethodDialog(
+    riderFiatMethods: List<String>,
+    onAcceptAnyway: () -> Unit,
+    onDecline: () -> Unit
+) {
+    AlertDialog(
+        onDismissRequest = onDecline,
+        icon = {
+            Icon(
+                Icons.Default.Warning,
+                contentDescription = null,
+                tint = MaterialTheme.colorScheme.error
+            )
+        },
+        title = { Text("No Common Payment Method") },
+        text = {
+            Column {
+                Text(
+                    "You and this rider don't share a payment method.",
+                    style = MaterialTheme.typography.bodyMedium
+                )
+                Spacer(modifier = Modifier.height(12.dp))
+                Text(
+                    "Rider's preferred methods:",
+                    style = MaterialTheme.typography.labelLarge
+                )
+                Spacer(modifier = Modifier.height(4.dp))
+                riderFiatMethods.take(3).forEach { method ->
+                    val displayName = com.ridestr.common.nostr.events.PaymentMethod.fromString(method)?.displayName ?: method
+                    Text(
+                        text = "  \u2022 $displayName",
+                        style = MaterialTheme.typography.bodyMedium
+                    )
+                }
+                if (riderFiatMethods.size > 3) {
+                    Text(
+                        text = "  + ${riderFiatMethods.size - 3} more",
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+            }
+        },
+        confirmButton = {
+            Button(onClick = onAcceptAnyway) {
+                Text("Accept Anyway")
+            }
+        },
+        dismissButton = {
+            TextButton(onClick = onDecline) {
+                Text("Decline Ride")
+            }
+        }
+    )
+}
+
+// ---------------------------------------------------------------------------
+// New composables extracted from inline code in stage-routing composables
+// ---------------------------------------------------------------------------
+
+/**
+ * Status card + toggle action buttons for the RoadFlare-Only mode.
+ * Extracted from the top of RoadflareOnlyContent.
+ *
+ * @param statusMessage  Subtitle shown inside the status card (currently hard-coded text).
+ * @param onGoFullyOnline  Callback for the "Go Fully Online" button.
+ * @param onGoOffline  Callback for the "Go Offline" button.
+ */
+@Composable
+fun AvailabilityControls(
+    onGoFullyOnline: () -> Unit,
+    onGoOffline: () -> Unit
+) {
+    Column(
+        modifier = Modifier.fillMaxWidth(),
+        verticalArrangement = Arrangement.spacedBy(12.dp)
+    ) {
+        // Status card
+        Card(
+            modifier = Modifier.fillMaxWidth(),
+            colors = CardDefaults.cardColors(
+                containerColor = MaterialTheme.colorScheme.primaryContainer.copy(alpha = 0.3f)
+            )
+        ) {
+            Row(
+                verticalAlignment = Alignment.CenterVertically,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(16.dp)
+            ) {
+                Icon(
+                    imageVector = Icons.Default.People,
+                    contentDescription = null,
+                    tint = MaterialTheme.colorScheme.primary,
+                    modifier = Modifier.size(32.dp)
+                )
+                Spacer(modifier = Modifier.width(12.dp))
+                Column(modifier = Modifier.weight(1f)) {
+                    Text(
+                        text = "RoadFlare Only",
+                        style = MaterialTheme.typography.titleMedium,
+                        color = MaterialTheme.colorScheme.primary
+                    )
+                    Text(
+                        text = "Broadcasting to followers, receiving RoadFlare requests",
+                        style = MaterialTheme.typography.bodySmall,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+            }
+        }
+
+        // Action buttons
+        Row(
+            horizontalArrangement = Arrangement.spacedBy(8.dp),
+            modifier = Modifier.fillMaxWidth()
+        ) {
+            Button(
+                onClick = onGoFullyOnline,
+                modifier = Modifier.weight(1f)
+            ) {
+                Icon(Icons.Default.Power, contentDescription = null, modifier = Modifier.size(18.dp))
+                Spacer(modifier = Modifier.width(4.dp))
+                Text("Go Fully Online")
+            }
+            OutlinedButton(
+                onClick = onGoOffline,
+                modifier = Modifier.weight(1f)
+            ) {
+                Icon(Icons.Default.PowerOff, contentDescription = null, modifier = Modifier.size(18.dp))
+                Spacer(modifier = Modifier.width(4.dp))
+                Text("Go Offline")
+            }
+        }
+    }
+}
+
+/**
+ * List of pending RoadFlare offers (or an empty-state card when none exist).
+ * Extracted from the lower section of RoadflareOnlyContent.
+ */
+@Composable
+fun RoadflareFollowerList(
+    pendingOffers: List<RideOfferData>,
+    isProcessingOffer: Boolean,
+    directOfferPickupRoutes: Map<String, RouteResult>,
+    directOfferRideRoutes: Map<String, RouteResult>,
+    displayCurrency: DisplayCurrency,
+    distanceUnit: DistanceUnit,
+    driverFiatMethods: List<String>,
+    onToggleCurrency: () -> Unit,
+    priceService: BitcoinPriceService,
+    onAcceptOffer: (RideOfferData) -> Unit,
+    onDeclineOffer: (RideOfferData) -> Unit,
+    onSetNoMatchWarning: (String) -> Unit
+) {
+    if (pendingOffers.isNotEmpty()) {
+        Text(
+            text = "RoadFlare Requests",
+            style = MaterialTheme.typography.titleMedium,
+            modifier = Modifier.padding(top = 8.dp)
+        )
+        pendingOffers.forEach { offer ->
+            RideOfferCard(
+                offer = offer,
+                pickupRoute = directOfferPickupRoutes[offer.eventId],
+                rideRoute = directOfferRideRoutes[offer.eventId],
+                isProcessing = isProcessingOffer,
+                onAccept = {
+                    // Intercept fiat offers with no common method (Issue #46)
+                    if (offer.isRoadflare && offer.paymentMethod != "cashu" &&
+                        offer.fiatPaymentMethods.isNotEmpty() &&
+                        com.ridestr.common.nostr.events.PaymentMethod.findBestCommonFiatMethod(
+                            offer.fiatPaymentMethods,
+                            effectiveRoadflareDriverMethods(driverFiatMethods)
+                        ) == null
+                    ) {
+                        onSetNoMatchWarning(offer.eventId)
+                    } else {
+                        onAcceptOffer(offer)
+                    }
+                },
+                onDecline = { onDeclineOffer(offer) },
+                displayCurrency = displayCurrency,
+                distanceUnit = distanceUnit,
+                onToggleCurrency = onToggleCurrency,
+                priceService = priceService,
+                driverFiatMethods = driverFiatMethods
+            )
+        }
+    } else {
+        Card(
+            modifier = Modifier.fillMaxWidth(),
+            colors = CardDefaults.cardColors(
+                containerColor = MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.5f)
+            )
+        ) {
+            Column(
+                horizontalAlignment = Alignment.CenterHorizontally,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(24.dp)
+            ) {
+                Icon(
+                    imageVector = Icons.Default.Sensors,
+                    contentDescription = null,
+                    modifier = Modifier.size(40.dp),
+                    tint = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+                Spacer(modifier = Modifier.height(8.dp))
+                Text(
+                    text = "Waiting for RoadFlare requests...",
+                    style = MaterialTheme.typography.bodyMedium,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant,
+                    textAlign = TextAlign.Center
+                )
+            }
+        }
+    }
+}
+
+/**
+ * Combined offers display: broadcast requests followed by direct offer cards.
+ * Extracted from the lower section of AvailableContent.
+ *
+ * When [totalRequests] is 0 an empty-state card is shown; otherwise a LazyColumn
+ * lists broadcast requests first, then direct offers.
+ */
+@Composable
+fun OfferInbox(
+    expandedSearch: Boolean,
+    freshBroadcastRequests: List<BroadcastRideOfferData>,
+    pendingOffers: List<RideOfferData>,
+    isProcessingOffer: Boolean,
+    pickupRoutes: Map<String, RouteResult>,
+    directOfferPickupRoutes: Map<String, RouteResult>,
+    directOfferRideRoutes: Map<String, RouteResult>,
+    displayCurrency: DisplayCurrency,
+    distanceUnit: DistanceUnit,
+    driverFiatMethods: List<String>,
+    onToggleCurrency: () -> Unit,
+    priceService: BitcoinPriceService,
+    onAcceptBroadcastRequest: (BroadcastRideOfferData) -> Unit,
+    onDeclineBroadcastRequest: (BroadcastRideOfferData) -> Unit,
+    onAcceptOffer: (RideOfferData) -> Unit,
+    onDeclineOffer: (RideOfferData) -> Unit,
+    onSetNoMatchWarning: (String) -> Unit,
+    onToggleExpandedSearch: () -> Unit
+) {
+    val totalRequests = freshBroadcastRequests.size + pendingOffers.size
+
+    // Search area toggle and ride requests header
+    Row(
+        modifier = Modifier.fillMaxWidth(),
+        horizontalArrangement = Arrangement.SpaceBetween,
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Text(
+            text = "Ride Requests ($totalRequests)",
+            style = MaterialTheme.typography.titleMedium,
+            fontWeight = FontWeight.Bold
+        )
+        FilterChip(
+            selected = expandedSearch,
+            onClick = onToggleExpandedSearch,
+            label = {
+                Text(if (expandedSearch) "Wide Area" else "Expand Search")
+            },
+            leadingIcon = {
+                Icon(
+                    imageVector = if (expandedSearch)
+                        Icons.Default.ZoomOutMap
+                    else
+                        Icons.Default.ZoomIn,
+                    contentDescription = null,
+                    modifier = Modifier.size(16.dp)
+                )
+            }
+        )
+    }
+
+    Spacer(modifier = Modifier.height(8.dp))
+
+    if (totalRequests == 0) {
+        Card(modifier = Modifier.fillMaxWidth()) {
+            Column(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(24.dp),
+                horizontalAlignment = Alignment.CenterHorizontally
+            ) {
+                Icon(
+                    Icons.Default.DirectionsCar,
+                    contentDescription = null,
+                    modifier = Modifier.size(40.dp),
+                    tint = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.5f)
+                )
+                Spacer(modifier = Modifier.height(8.dp))
+                Text(
+                    text = "Waiting for ride requests...",
+                    textAlign = TextAlign.Center,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant
+                )
+                Text(
+                    text = "Requests from nearby riders will appear here",
+                    style = MaterialTheme.typography.bodySmall,
+                    textAlign = TextAlign.Center,
+                    color = MaterialTheme.colorScheme.onSurfaceVariant.copy(alpha = 0.7f)
+                )
+            }
+        }
+    } else {
+        LazyColumn(
+            modifier = Modifier.weight(1f),
+            verticalArrangement = Arrangement.spacedBy(8.dp)
+        ) {
+            // Show broadcast requests first (sorted by fare, highest first)
+            // Using freshBroadcastRequests which filters out stale requests
+            items(freshBroadcastRequests) { request ->
+                BroadcastRideRequestCard(
+                    request = request,
+                    pickupRoute = pickupRoutes[request.eventId],
+                    isProcessing = isProcessingOffer,
+                    onAccept = { onAcceptBroadcastRequest(request) },
+                    onDecline = { onDeclineBroadcastRequest(request) },
+                    displayCurrency = displayCurrency,
+                    distanceUnit = distanceUnit,
+                    onToggleCurrency = onToggleCurrency,
+                    priceService = priceService
+                )
+            }
+
+            // Show direct offers (legacy/advanced) if any
+            if (pendingOffers.isNotEmpty()) {
+                item {
+                    Spacer(modifier = Modifier.height(8.dp))
+                    Text(
+                        text = "Direct Requests (${pendingOffers.size})",
+                        style = MaterialTheme.typography.labelMedium,
+                        color = MaterialTheme.colorScheme.onSurfaceVariant
+                    )
+                }
+                items(pendingOffers) { offer ->
+                    RideOfferCard(
+                        offer = offer,
+                        pickupRoute = directOfferPickupRoutes[offer.eventId],
+                        rideRoute = directOfferRideRoutes[offer.eventId],
+                        isProcessing = isProcessingOffer,
+                        onAccept = {
+                            // Intercept fiat offers with no common method (Issue #46)
+                            if (offer.isRoadflare && offer.paymentMethod != "cashu" &&
+                                offer.fiatPaymentMethods.isNotEmpty() &&
+                                com.ridestr.common.nostr.events.PaymentMethod.findBestCommonFiatMethod(
+                                    offer.fiatPaymentMethods,
+                                    effectiveRoadflareDriverMethods(driverFiatMethods)
+                                ) == null
+                            ) {
+                                onSetNoMatchWarning(offer.eventId)
+                            } else {
+                                onAcceptOffer(offer)
+                            }
+                        },
+                        onDecline = { onDeclineOffer(offer) },
+                        displayCurrency = displayCurrency,
+                        distanceUnit = distanceUnit,
+                        onToggleCurrency = onToggleCurrency,
+                        priceService = priceService,
+                        driverFiatMethods = driverFiatMethods
+                    )
+                }
+            }
+        }
+    }
+}
