+++ roadflare-rider/src/main/java/com/roadflare/rider/ui/screens/DriverNetworkTab.kt
@@ -1,52 +1,28 @@
 package com.roadflare.rider.ui.screens
 
-import android.content.Intent
 import android.util.Log
 import com.roadflare.rider.BuildConfig
 import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.*
 import androidx.compose.material3.*
-import com.ridestr.common.nostr.events.PaymentMethod
 import androidx.compose.material3.pulltorefresh.PullToRefreshBox
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
-import com.roadflare.rider.R
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import com.ridestr.common.data.FollowedDriversRepository
 import com.ridestr.common.nostr.NostrService
 import com.ridestr.common.nostr.events.FollowedDriver
 import com.ridestr.common.nostr.events.Location
-import com.ridestr.common.nostr.events.RoadflareLocationEvent
-import com.vitorpamplona.quartz.nip01Core.core.hexToByteArray
-import com.vitorpamplona.quartz.nip19Bech32.toNpub
-import java.text.SimpleDateFormat
-import java.util.*
+import com.roadflare.rider.ui.screens.components.DriverLocationState
+import com.roadflare.rider.ui.screens.components.EmptyDriversState
+import com.roadflare.rider.ui.screens.components.FollowedDriverList
+import com.roadflare.rider.ui.screens.components.RoadflarePaymentMethodsDialog
 
 private const val TAG = "DriverNetworkTab"
-
-/**
- * Represents a driver's real-time location and status from RoadFlare broadcast.
- */
-data class DriverLocationState(
-    val pubkey: String,
-    val lat: Double,
-    val lon: Double,
-    val status: String, // "online", "on_ride", "offline"
-    val timestamp: Long,
-    val keyVersion: Int
-)
-
-// Fare calculation handled by RoadflareFarePolicy in :common (per-driver, config-aware)
 
 /**
  * RoadFlare tab showing rider's favorite drivers list.
@@ -67,19 +43,16 @@ fun DriverNetworkTab(
     riderLocation: Location? = null,
     onAddDriver: () -> Unit = {},
     onDriverClick: (FollowedDriver) -> Unit = {},
-    onDriverRemoved: () -> Unit = {},  // Called after driver removed - triggers Nostr backup
+    onDriverRemoved: () -> Unit = {},
     modifier: Modifier = Modifier
 ) {
-    val context = LocalContext.current
     val scope = rememberCoroutineScope()
     val drivers by followedDriversRepository.drivers.collectAsState()
     val driverNames by followedDriversRepository.driverNames.collectAsState()
     val cachedLocations by followedDriversRepository.driverLocations.collectAsState()
 
-    // Payment methods dialog state
     var showPaymentMethodsDialog by remember { mutableStateOf(false) }
 
-    // DEBUG: Log rider's followed drivers state (debug builds only)
     LaunchedEffect(drivers) {
         if (BuildConfig.DEBUG) {
             Log.d(TAG, "=== RIDER ROADFLARE STATE ===")
@@ -96,22 +69,13 @@ fun DriverNetworkTab(
         }
     }
 
-    // NOTE: Location subscriptions and profile fetching are now managed by
-    // RoadflareDriverPresenceCoordinator in the ViewModel (activity-scoped).
-    // This tab reads from repository StateFlows only.
-
     var showRemoveDialog by remember { mutableStateOf<FollowedDriver?>(null) }
 
-    // Track which drivers have stale keys (need key refresh from driver)
     val staleKeyDrivers = remember { mutableStateMapOf<String, Boolean>() }
-
-    // Track when we last requested a key refresh for each driver (rate limiting)
     val keyRefreshRequests = remember { mutableStateMapOf<String, Long>() }
 
-    // Pull to refresh state
     var isRefreshing by remember { mutableStateOf(false) }
 
-    // Check for stale keys on refresh - also auto-request key refresh
     fun checkStaleKeys() {
         if (nostrService == null) return
         scope.launch {
@@ -119,15 +83,12 @@ fun DriverNetworkTab(
                 val storedKey = driver.roadflareKey
                 val storedKeyUpdatedAt = storedKey?.keyUpdatedAt ?: 0L
 
-                // If we have no key yet, request a fresh key share (rate-limited).
                 if (storedKey == null) {
                     val lastRefreshRequest = keyRefreshRequests[driver.pubkey] ?: 0L
                     val now = System.currentTimeMillis()
                     val oneHourMs = 3600_000L
-
                     if (now - lastRefreshRequest > oneHourMs) {
                         keyRefreshRequests[driver.pubkey] = now
-
                         val eventId = nostrService.publishRoadflareKeyAck(
                             driverPubKey = driver.pubkey,
                             keyVersion = 0,
@@ -142,32 +103,26 @@ fun DriverNetworkTab(
                     } else {
                         Log.d(TAG, "Skipping key refresh (no key) for ${driver.pubkey.take(8)} - rate limited (last request ${(now - lastRefreshRequest) / 1000}s ago)")
                     }
-
                     staleKeyDrivers[driver.pubkey] = false
                     return@forEach
                 }
 
                 if (storedKeyUpdatedAt > 0) {
-                    // Fetch driver's current key_updated_at from their Kind 30012
                     val currentKeyUpdatedAt = nostrService.fetchDriverKeyUpdatedAt(driver.pubkey)
                     if (currentKeyUpdatedAt != null && currentKeyUpdatedAt > storedKeyUpdatedAt) {
                         Log.d(TAG, "Stale key detected for driver ${driver.pubkey.take(8)}: stored=$storedKeyUpdatedAt, current=$currentKeyUpdatedAt")
                         staleKeyDrivers[driver.pubkey] = true
 
-                        // Auto-request key refresh (rate-limited to once per hour per driver)
                         val lastRefreshRequest = keyRefreshRequests[driver.pubkey] ?: 0L
                         val now = System.currentTimeMillis()
-                        val oneHourMs = 3600_000L  // 1 hour in milliseconds
-
+                        val oneHourMs = 3600_000L
                         if (now - lastRefreshRequest > oneHourMs) {
                             keyRefreshRequests[driver.pubkey] = now
-
-                            // Send Kind 3188 with status="stale" to request key refresh
                             val eventId = nostrService.publishRoadflareKeyAck(
                                 driverPubKey = driver.pubkey,
                                 keyVersion = driver.roadflareKey?.version ?: 0,
                                 keyUpdatedAt = storedKeyUpdatedAt,
-                                status = "stale"  // Indicates refresh request
+                                status = "stale"
                             )
                             if (eventId != null) {
                                 Log.d(TAG, "Requested key refresh for ${driver.pubkey.take(8)}: eventId=${eventId.take(8)}")
@@ -185,10 +140,9 @@ fun DriverNetworkTab(
         }
     }
 
-    // Check stale keys on initial load
     LaunchedEffect(drivers) {
         if (nostrService != null && drivers.isNotEmpty()) {
-            delay(3000) // Wait for connection
+            delay(3000)
             checkStaleKeys()
         }
     }
@@ -196,16 +150,13 @@ fun DriverNetworkTab(
     fun onRefresh() {
         scope.launch {
             isRefreshing = true
-            // Refresh driver locations by re-subscribing
             followedDriversRepository.clearDriverLocations()
-            // Check for stale keys
             checkStaleKeys()
-            delay(500) // Small delay to show refresh indicator
+            delay(500)
             isRefreshing = false
         }
     }
 
-    // Remove driver confirmation dialog
     showRemoveDialog?.let { driver ->
         val dialogDriverName = driverNames[driver.pubkey] ?: driver.pubkey.take(8) + "..."
         AlertDialog(
@@ -218,10 +169,8 @@ fun DriverNetworkTab(
             confirmButton = {
                 TextButton(
                     onClick = {
-                        // Remove driver from favorites - driver will see via p-tag query
                         followedDriversRepository.removeDriver(driver.pubkey)
                         showRemoveDialog = null
-                        // Trigger Nostr backup to publish updated Kind 30011 (without p-tag)
                         onDriverRemoved()
                     }
                 ) {
@@ -236,13 +185,26 @@ fun DriverNetworkTab(
         )
     }
 
-    // Payment methods dialog
     if (showPaymentMethodsDialog) {
         RoadflarePaymentMethodsDialog(
             currentMethods = roadflarePaymentMethods,
             onSave = onSetRoadflarePaymentMethods,
             onDismiss = { showPaymentMethodsDialog = false }
         )
+    }
+
+    // Convert cached locations to DriverLocationState for presentation
+    val locationStates = remember(cachedLocations) {
+        cachedLocations.mapValues { (pubkey, cached) ->
+            DriverLocationState(
+                pubkey = pubkey,
+                lat = cached.lat,
+                lon = cached.lon,
+                status = cached.status,
+                timestamp = cached.timestamp,
+                keyVersion = cached.keyVersion
+            )
+        }
     }
 
     Scaffold(
@@ -252,16 +214,12 @@ fun DriverNetworkTab(
                     horizontalAlignment = Alignment.End,
                     verticalArrangement = Arrangement.spacedBy(12.dp)
                 ) {
-                    // Payment methods FAB
                     SmallFloatingActionButton(
                         onClick = { showPaymentMethodsDialog = true }
                     ) {
                         Icon(Icons.Default.Payment, contentDescription = "Payment Methods")
                     }
-                    // Add driver FAB
-                    FloatingActionButton(
-                        onClick = onAddDriver
-                    ) {
+                    FloatingActionButton(onClick = onAddDriver) {
                         Icon(Icons.Default.PersonAdd, contentDescription = "Add Driver")
                     }
                 }
@@ -277,431 +235,21 @@ fun DriverNetworkTab(
                 .padding(innerPadding)
         ) {
             if (drivers.isEmpty()) {
-                // Empty state
                 EmptyDriversState(
                     onAddDriver = onAddDriver,
                     modifier = Modifier.fillMaxSize()
                 )
             } else {
-                // Drivers list
-                LazyColumn(
-                    contentPadding = PaddingValues(
-                        start = 16.dp,
-                        end = 16.dp,
-                        top = 8.dp,
-                        bottom = 100.dp // Space for FAB
-                    ),
-                    verticalArrangement = Arrangement.spacedBy(8.dp),
-                    modifier = Modifier.fillMaxSize()
-                ) {
-                    item {
-                        Row(
-                            verticalAlignment = Alignment.CenterVertically,
-                            modifier = Modifier.fillMaxWidth().padding(bottom = 8.dp)
-                        ) {
-                            Text(
-                                text = "Favorite Drivers",
-                                style = MaterialTheme.typography.headlineSmall,
-                                modifier = Modifier.weight(1f)
-                            )
-                            if (drivers.size > 1) {
-                                IconButton(onClick = {
-                                    val shareText = drivers.joinToString("\n") { d ->
-                                        val name = driverNames[d.pubkey]
-                                        val npub = d.pubkey.hexToByteArray().toNpub()
-                                        if (name != null) "$name: nostr:$npub" else "nostr:$npub"
-                                    }
-                                    val intent = Intent(Intent.ACTION_SEND).apply {
-                                        type = "text/plain"
-                                        putExtra(Intent.EXTRA_TEXT, shareText)
-                                    }
-                                    context.startActivity(Intent.createChooser(intent, "Share driver list"))
-                                }) {
-                                    Icon(
-                                        Icons.Default.Share,
-                                        contentDescription = "Share all drivers"
-                                    )
-                                }
-                            }
-                        }
-                    }
-
-                    items(drivers, key = { it.pubkey }) { driver ->
-                        // Convert CachedDriverLocation to DriverLocationState for display
-                        val locationState = cachedLocations[driver.pubkey]?.let { cached ->
-                            DriverLocationState(
-                                pubkey = driver.pubkey,
-                                lat = cached.lat,
-                                lon = cached.lon,
-                                status = cached.status,
-                                timestamp = cached.timestamp,
-                                keyVersion = cached.keyVersion
-                            )
-                        }
-                        DriverCard(
-                            driver = driver,
-                            driverName = driverNames[driver.pubkey],
-                            locationState = locationState,
-                            riderLocation = riderLocation,
-                            hasStaleKey = staleKeyDrivers[driver.pubkey] == true,
-                            onClick = { onDriverClick(driver) },
-                            onRemove = { showRemoveDialog = driver },
-                            onShare = {
-                                val npub = driver.pubkey.hexToByteArray().toNpub()
-                                val intent = Intent(Intent.ACTION_SEND).apply {
-                                    type = "text/plain"
-                                    putExtra(Intent.EXTRA_TEXT, "nostr:$npub")
-                                }
-                                context.startActivity(Intent.createChooser(intent, "Share driver"))
-                            }
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-/**
- * Empty state when rider has no followed drivers.
- */
-@Composable
-private fun EmptyDriversState(
-    onAddDriver: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-        modifier = modifier.padding(32.dp)
-    ) {
-        Icon(
-            painter = painterResource(R.drawable.ic_roadflare),
-            contentDescription = null,
-            modifier = Modifier.size(140.dp),
-            tint = MaterialTheme.colorScheme.primary
-        )
-
-        Spacer(modifier = Modifier.height(24.dp))
-
-        Text(
-            text = "Build Your Network",
-            style = MaterialTheme.typography.headlineMedium
-        )
-
-        Spacer(modifier = Modifier.height(8.dp))
-
-        Text(
-            text = "Add drivers you've met and trust to your personal RoadFlare network. Send out a RoadFlare and it requests a ride straight from your trusted circle of favorite drivers.",
-            style = MaterialTheme.typography.bodyMedium,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            modifier = Modifier.padding(horizontal = 16.dp)
-        )
-
-        Spacer(modifier = Modifier.height(32.dp))
-
-        Button(
-            onClick = onAddDriver,
-            modifier = Modifier.fillMaxWidth(0.7f)
-        ) {
-            Icon(Icons.Default.PersonAdd, contentDescription = null)
-            Spacer(modifier = Modifier.width(8.dp))
-            Text("Add Your First Driver")
-        }
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        Text(
-            text = "Scan a driver's QR code or enter their Nostr pubkey",
-            style = MaterialTheme.typography.bodySmall,
-            color = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-    }
-}
-
-/**
- * Card displaying a single followed driver.
- *
- * @param driver The driver data (pubkey, note, key)
- * @param driverName Display name fetched from Nostr profile (or null if not yet loaded)
- */
-@Composable
-private fun DriverCard(
-    driver: FollowedDriver,
-    driverName: String?,
-    locationState: DriverLocationState?,
-    riderLocation: Location?,
-    hasStaleKey: Boolean = false,
-    onClick: () -> Unit,
-    onRemove: () -> Unit,
-    onShare: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val dateFormat = remember { SimpleDateFormat("MMM d, yyyy", Locale.getDefault()) }
-    val displayName = driverName ?: driver.pubkey.take(8) + "..."
-
-    // Calculate distance to driver if both locations available
-    val distanceMiles = if (riderLocation != null && locationState != null &&
-        locationState.status == RoadflareLocationEvent.Status.ONLINE) {
-        val driverLoc = Location(locationState.lat, locationState.lon)
-        val distanceKm = riderLocation.distanceToKm(driverLoc)
-        distanceKm * 0.621371
-    } else null
-
-    Card(
-        onClick = onClick,
-        modifier = modifier.fillMaxWidth()
-    ) {
-        Row(
-            verticalAlignment = Alignment.CenterVertically,
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp)
-        ) {
-            // Driver avatar placeholder
-            Surface(
-                shape = MaterialTheme.shapes.medium,
-                color = MaterialTheme.colorScheme.primaryContainer,
-                modifier = Modifier.size(48.dp)
-            ) {
-                Box(contentAlignment = Alignment.Center) {
-                    Text(
-                        text = displayName.firstOrNull()?.uppercase() ?: "?",
-                        style = MaterialTheme.typography.titleLarge,
-                        color = MaterialTheme.colorScheme.onPrimaryContainer
-                    )
-                }
-            }
-
-            Spacer(modifier = Modifier.width(16.dp))
-
-            // Driver info
-            Column(modifier = Modifier.weight(1f)) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Text(
-                        text = displayName,
-                        style = MaterialTheme.typography.titleMedium,
-                        maxLines = 1,
-                        overflow = TextOverflow.Ellipsis,
-                        modifier = Modifier.weight(1f, fill = false)
-                    )
-
-                    Spacer(modifier = Modifier.width(8.dp))
-
-                    // Real-time status from location broadcast
-                    DriverStatusBadge(
-                        hasKey = driver.roadflareKey != null,
-                        locationState = locationState,
-                        hasStaleKey = hasStaleKey
-                    )
-                }
-
-                // Show driver note or distance
-                if (distanceMiles != null) {
-                    Text(
-                        text = formatDistance(distanceMiles),
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = MaterialTheme.colorScheme.primary,
-                        maxLines = 1,
-                        overflow = TextOverflow.Ellipsis
-                    )
-                } else if (driver.note.isNotEmpty()) {
-                    Text(
-                        text = driver.note,
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        maxLines = 1,
-                        overflow = TextOverflow.Ellipsis
-                    )
-                }
-
-                // Show last seen time or "Added" date
-                val statusText = if (locationState != null) {
-                    val lastSeen = formatLastSeen(locationState.timestamp)
-                    "Last seen $lastSeen"
-                } else {
-                    "Added ${dateFormat.format(Date(driver.addedAt * 1000))}"
-                }
-                Text(
-                    text = statusText,
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-
-            // Share button
-            IconButton(onClick = onShare) {
-                Icon(
-                    Icons.Default.Share,
-                    contentDescription = "Share",
-                    tint = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-
-            // Remove button
-            IconButton(onClick = onRemove) {
-                Icon(
-                    Icons.Default.Close,
-                    contentDescription = "Remove",
-                    tint = MaterialTheme.colorScheme.onSurfaceVariant
+                FollowedDriverList(
+                    drivers = drivers,
+                    driverNames = driverNames,
+                    locationStates = locationStates,
+                    riderLocation = riderLocation,
+                    staleKeyDrivers = staleKeyDrivers,
+                    onDriverClick = onDriverClick,
+                    onRemove = { showRemoveDialog = it }
                 )
             }
         }
     }
-}
-
-/**
- * Format distance in a user-friendly way.
- */
-private fun formatDistance(miles: Double): String {
-    return when {
-        miles < 0.1 -> "Very close"
-        miles < 1.0 -> String.format("%.1f mi away", miles)
-        miles < 10.0 -> String.format("%.1f mi away", miles)
-        else -> String.format("%.0f mi away", miles)
-    }
-}
-
-/**
- * Status badge showing driver's real-time status.
- *
- * Status mapping:
- * - online -> "Available" (Green)
- * - on_ride -> "On Ride" (Yellow/Orange)
- * - offline -> "Offline" (Gray)
- * - No location / stale -> "Offline" or "Pending" (Gray)
- */
-@Composable
-private fun DriverStatusBadge(
-    hasKey: Boolean,
-    locationState: DriverLocationState?,
-    hasStaleKey: Boolean = false,
-    modifier: Modifier = Modifier
-) {
-    // Check if location is stale (older than 5 minutes)
-    val isStale = locationState?.let {
-        val now = System.currentTimeMillis() / 1000
-        (now - it.timestamp) > 300 // 5 minutes
-    } ?: true
-
-    // Priority order: stale key -> no key -> explicit OFFLINE -> stale data -> status-based
-    // IMPORTANT: Check explicit OFFLINE status BEFORE staleness to ensure immediate offline detection
-    val (color, text) = when {
-        hasStaleKey -> MaterialTheme.colorScheme.error to "Key Outdated"
-        !hasKey -> MaterialTheme.colorScheme.outline to "Pending"
-        locationState == null -> MaterialTheme.colorScheme.outline to "Offline"
-        locationState.status == RoadflareLocationEvent.Status.OFFLINE ->
-            MaterialTheme.colorScheme.outline to "Offline"  // Explicit offline FIRST (before staleness)
-        isStale -> MaterialTheme.colorScheme.outline to "Offline"  // Stale fallback
-        locationState.status == RoadflareLocationEvent.Status.ONLINE ->
-            MaterialTheme.colorScheme.primary to "Available"
-        locationState.status == RoadflareLocationEvent.Status.ON_RIDE ->
-            MaterialTheme.colorScheme.tertiary to "On Ride"
-        else -> MaterialTheme.colorScheme.outline to "Offline"
-    }
-
-    Surface(
-        shape = MaterialTheme.shapes.small,
-        color = color.copy(alpha = 0.1f),
-        modifier = modifier
-    ) {
-        Text(
-            text = text,
-            style = MaterialTheme.typography.labelSmall,
-            color = color,
-            modifier = Modifier.padding(horizontal = 8.dp, vertical = 2.dp)
-        )
-    }
-}
-
-/**
- * Format timestamp as "just now", "X min ago", etc.
- */
-private fun formatLastSeen(timestamp: Long): String {
-    val now = System.currentTimeMillis() / 1000
-    val diff = now - timestamp
-
-    return when {
-        diff < 60 -> "just now"
-        diff < 3600 -> "${diff / 60} min ago"
-        diff < 86400 -> "${diff / 3600} hr ago"
-        else -> "${diff / 86400} days ago"
-    }
-}
-
-// decryptRoadflareLocation moved to RoadflareDriverPresenceCoordinator in :common
-
-/**
- * Dialog for selecting RoadFlare alternate payment methods.
- * These are fiat methods riders can offer to personal RoadFlare drivers.
- */
-@Composable
-fun RoadflarePaymentMethodsDialog(
-    currentMethods: List<String>,
-    onSave: (List<String>) -> Unit,
-    onDismiss: () -> Unit
-) {
-    // Local state for Save/Cancel pattern
-    var localMethods by remember(currentMethods) { mutableStateOf(currentMethods) }
-
-    // All known methods + any unknown stored strings
-    val allMethods = remember(localMethods) {
-        val known = PaymentMethod.ROADFLARE_ALTERNATE_METHODS.map { it.value }
-        (known + localMethods.filter { it !in known }).distinct()
-    }
-
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        icon = {
-            Icon(
-                Icons.Default.Payment,
-                contentDescription = null,
-                tint = MaterialTheme.colorScheme.primary
-            )
-        },
-        title = { Text("Payment Methods") },
-        text = {
-            Column {
-                Text(
-                    "For your personal RoadFlare drivers, select which " +
-                    "payment methods you can offer.",
-                    style = MaterialTheme.typography.bodyMedium
-                )
-                Spacer(modifier = Modifier.height(12.dp))
-                Text(
-                    "Select and drag to set priority order:",
-                    style = MaterialTheme.typography.labelLarge
-                )
-                Spacer(modifier = Modifier.height(8.dp))
-                com.ridestr.common.ui.components.ReorderablePaymentMethodList(
-                    allMethods = allMethods,
-                    enabledMethods = localMethods,
-                    onOrderChanged = { reordered -> localMethods = reordered },
-                    onMethodToggled = { method, enabled ->
-                        localMethods = if (enabled) {
-                            localMethods + method
-                        } else {
-                            localMethods - method
-                        }
-                    },
-                    modifier = Modifier.heightIn(max = 350.dp)
-                )
-            }
-        },
-        confirmButton = {
-            Button(onClick = {
-                onSave(localMethods)
-                onDismiss()
-            }) {
-                Text("Save")
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Cancel")
-            }
-        }
-    )
 }
